ETSWORK-16638 refactor hooks
