+++ examples/react/json-based-notification-renderer-with-css/src/components/Notifications.jsx
@@ -17,13 +17,12 @@
 import React from 'react'
 import Notification from './Notification'
 import InfiniteScroll from 'react-infinite-scroll-component'
-import useNotifications from './hooks/useNotifications'
+import useFetchNotifications from './hooks/useFetchNotifications'
 import useDeleteNotification from './hooks/useDeleteNotification'
 import useActionNotification from './hooks/useActionNotification'
 
 export default function Notifications() {
-  // Hooks to handle GraphQL queries
-  const [notifications, loading, error, hasNextPage, fetchMoreNotifications] = useNotifications()
+  const [notifications, loading, error, hasNextPage, fetchMore] = useFetchNotifications()
   const [deleteNotification] = useDeleteNotification()
   const [actionNotification] = useActionNotification()
 
@@ -57,7 +56,7 @@ export default function Notifications() {
   return (
     <InfiniteScroll
       dataLength={notifications.length}
-      next={fetchMoreNotifications}
+      next={fetchMore}
       hasMore={hasNextPage}
       loader={<h4>Loading more notifications...</h4>}
       endMessage={
