+++ examples/react/json-based-notification-renderer-with-custom-components/src/components/hooks/useActionNotification.jsx
@@ -1,6 +1,6 @@
 import { useMutation, gql } from '@apollo/client'
-import { GET_NOTIFICATIONS_QUERY } from './useFetchNotifications'
 import { useApolloClient } from '@apollo/client'
+import deleteNotificationFromCache from './utils/deleteNotificationFromCache'
 
 // GraphQL Mutation to action on a notification
 const ACTION_NOTIFICATION_MUTATION = gql`
@@ -14,40 +14,7 @@ export default function useActionNotification() {
   const [actionNotificationFunction, { loading, error }] = useMutation(ACTION_NOTIFICATION_MUTATION, {
     // Fired on optimistic response, and on actual response
     update(cache, mutationResult) {
-      const queryOptions = {
-        query: GET_NOTIFICATIONS_QUERY,
-        variables: {
-          spaceId: process.env.REACT_APP_SPACE_ID,
-          input: {
-            location: 'TOKNOW',
-            filter: {
-              orderBy: 'DATE',
-            },
-          },
-        },
-      }
-
-      // Query the cache directly to get existing notifications
-      const cachedData = cache.readQuery(queryOptions)
-
-      // Grab id of notification that was deleted
-      const deletedId = mutationResult?.data?.actionNotification
-
-      // Re-write the cache without the deleted notification
-      const data = {
-        me: {
-          ...cachedData.me,
-          space: {
-            ...cachedData.me.space,
-            notifications: {
-              ...cachedData.me.space.notifications,
-              edges: cachedData.me.space.notifications.edges.filter((edge) => edge.node.id !== deletedId),
-              pageInfo: cachedData.me.space.notifications.pageInfo,
-            },
-          },
-        },
-      }
-      cache.writeQuery({ ...queryOptions, data })
+      deleteNotificationFromCache(cache, mutationResult)
     },
     // On successful delete, clean up cache
     onCompleted() {
